Remove PV checks and fetching
