+++ controllers/storage.go
@@ -2,7 +2,6 @@ package controllers
 
 import (
 	"context"
-	"fmt"
 	trustyaiopendatahubiov1alpha1 "github.com/ruivieira/trustyai-service-operator/api/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -13,36 +12,21 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-func (r *TrustyAIServiceReconciler) ensurePV(ctx context.Context, instance *trustyaiopendatahubiov1alpha1.TrustyAIService) (*corev1.PersistentVolume, error) {
-	// Extract the PV name from the instance spec
-	pvName := instance.Spec.Storage.PV
-
-	// Create a PV object
-	pv := &corev1.PersistentVolume{}
-
-	// Try to get the PV
-	err := r.Get(ctx, types.NamespacedName{Name: pvName}, pv)
-	if err != nil {
-		// If an error occurs while getting the PV, return the error
-		if apierrors.IsNotFound(err) {
-			return nil, fmt.Errorf("PersistentVolume %s not found", pvName)
-		}
-		return nil, err
-	}
-
-	// If the PV exists, return its reference
-	return pv, nil
+func generatePVCName(instance *trustyaiopendatahubiov1alpha1.TrustyAIService) string {
+	return instance.Name + "-pvc"
 }
 
-func (r *TrustyAIServiceReconciler) ensurePVC(ctx context.Context, instance *trustyaiopendatahubiov1alpha1.TrustyAIService, pv *corev1.PersistentVolume) error {
+func (r *TrustyAIServiceReconciler) ensurePVC(ctx context.Context, instance *trustyaiopendatahubiov1alpha1.TrustyAIService) error {
+	pvcName := generatePVCName(instance)
+
 	pvc := &corev1.PersistentVolumeClaim{}
 
-	err := r.Get(ctx, types.NamespacedName{Name: defaultPvcName, Namespace: instance.Namespace}, pvc)
+	err := r.Get(ctx, types.NamespacedName{Name: pvcName, Namespace: instance.Namespace}, pvc)
 	if err != nil {
 		if apierrors.IsNotFound(err) {
 			log.FromContext(ctx).Info("PVC not found. Creating.")
 			// The PVC doesn't exist, so we need to create it
-			return r.createPVC(ctx, instance, pv)
+			return r.createPVC(ctx, instance)
 		}
 		return err
 	}
@@ -50,14 +34,12 @@ func (r *TrustyAIServiceReconciler) ensurePVC(ctx context.Context, instance *tru
 	return nil
 }
 
-func (r *TrustyAIServiceReconciler) createPVC(ctx context.Context, instance *trustyaiopendatahubiov1alpha1.TrustyAIService, pv *corev1.PersistentVolume) error {
-
-	// Extract the storage class from the PV
-	storageClass := pv.Spec.StorageClassName
+func (r *TrustyAIServiceReconciler) createPVC(ctx context.Context, instance *trustyaiopendatahubiov1alpha1.TrustyAIService) error {
+	pvcName := generatePVCName(instance)
 
 	pvc := &corev1.PersistentVolumeClaim{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      defaultPvcName,
+			Name:      pvcName,
 			Namespace: instance.Namespace,
 		},
 		Spec: corev1.PersistentVolumeClaimSpec{
@@ -69,8 +51,6 @@ func (r *TrustyAIServiceReconciler) createPVC(ctx context.Context, instance *tru
 					corev1.ResourceStorage: resource.MustParse(instance.Spec.Storage.Size),
 				},
 			},
-			StorageClassName: &storageClass,
-			VolumeMode:       pv.Spec.VolumeMode,
 		},
 	}
